Major refactor of this project
I've change this project to be more modular and easier to develop.
- Two new classes: `PermanentRequest` and `PermanentAPI`
- Run this as a module instead of a script
- Format the whole thing with black, and add black as a dev requirement
- Add some basic CI tests (more to come), since this will soon be part of
  our deployment pipeline
- Update the existing workflows
