+++ .github/workflows/run-dev.yml
@@ -23,4 +23,4 @@ jobs:
         - name: Unarchive test files
           run: unzip one-of-each-type.zip
         - name: Run the test
-          run: python run_test.py dev files/
+          run: python -m permanent_upload dev files/
