+++ .github/workflows/test.yml
@@ -1,0 +1,24 @@
+name: Test
+
+on:
+  push:
+    branches:    
+      - main
+  pull_request:
+
+jobs:
+    execute:
+        runs-on: ubuntu-latest
+        steps:
+        - uses: actions/checkout@v1
+        - uses: actions/setup-python@v2
+          with:
+            python-version: '3.x'
+        - name: Install dependencies
+          run: pip install -r requirements-dev.txt
+        - name: Run formatter 
+          run: black --check permanent_upload
+        - name: Run package help
+          run: python3 -m permanent-upload -h
+
+
